requirements.txt
discord.py>=2.3.2
aiogram>=3.9.0